refactor(lib): clean up some missed path updates, update snapshots
