+++ src/lib/vec/types/index.ts
@@ -1,8 +1,8 @@
-import Vector from '.';
+import Vector from '..';
 
-import { Vec2D1, Vec2D2, Vec2D3, Vec2D4 } from './types/vec2';
-import { Vec3D1, Vec3D2, Vec3D3, Vec3D4 } from './types/vec3';
-import { Vec4D1, Vec4D2, Vec4D3, Vec4D4 } from './types/vec4';
+import { Vec2D1, Vec2D2, Vec2D3, Vec2D4 } from './vec2';
+import { Vec3D1, Vec3D2, Vec3D3, Vec3D4 } from './vec3';
+import { Vec4D1, Vec4D2, Vec4D3, Vec4D4 } from './vec4';
 
 export type Component = number | number[] | Vector;
 export type Components = Component[];
