+++ src/lib/vec/types/vec3/index.ts
@@ -1,6 +1,6 @@
-import { Rgb2, Rgb3, Rgb4 } from './vec3/rgb';
-import { Stp2, Stp3, Stp4 } from './vec3/stp';
-import { Xyz2, Xyz3, Xyz4 } from './vec3/xyz';
+import { Rgb2, Rgb3, Rgb4 } from './rgb';
+import { Stp2, Stp3, Stp4 } from './stp';
+import { Xyz2, Xyz3, Xyz4 } from './xyz';
 
 export type Vec3D1 = 'b' | 'p' | 'z';
 export type Vec3D2 = Rgb2 | Stp2 | Xyz2;
