+++ src/lib/vec/types/vec4/index.ts
@@ -1,6 +1,6 @@
-import { Rgba2, Rgba3, Rgba4 } from './vec4/rgba';
-import { Stpq2, Stpq3, Stpq4 } from './vec4/stpq';
-import { Xyzw2, Xyzw3, Xyzw4 } from './vec4/xyzw';
+import { Rgba2, Rgba3, Rgba4 } from './rgba';
+import { Stpq2, Stpq3, Stpq4 } from './stpq';
+import { Xyzw2, Xyzw3, Xyzw4 } from './xyzw';
 
 export type Vec4D1 = 'a' | 'q' | 'w';
 export type Vec4D2 = Rgba2 | Stpq2 | Xyzw2;
